requirements.txt
flask
Pillow
numpy
tensorflow
